src/agent-python/requirements.txt
livekit-agents[silero]~=1.3
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-google
python-dotenv
